Adds suport for object templates
Adds the ability to define a `template` property in the schema of properties fo type **object** that will be merged with the object passed in the configuration.

This allows scenarios where only a subset of the properties of the object will be defined in the configuration but we want to maintain the remanining properties defaults. The name is different from `default` to avoid changes in the current behavior.